requirements.txt
eth_abi==4.0.0
eth_account==0.8.0
eth_utils==2.1.0
Requests==2.31.0
semantic_version==2.10.0
tqdm==4.65.0
web3==6.11.1
pytest==7.4.0
eth-brownie
